Fix off-by-one bug in dbIter seqNum handling
The visible sequence number is less than or equal to the log sequence
number. dbIter was handling the visible sequence number incorrectly,
allowing entries with sequence numbers equal to the visible sequence
number to be seen, while only entries that are less than the visible
sequence number should be visible.