Fix the bugs exposed by TestRangeDelCompactionTruncation
levelIter needs to properly nil out levelIter.rangeDelIter when
switching to a new table. This is required for correctness by
mergingIter which needs to stop considering range tombstones for a table
once we've iterated past the table.

Fixed a similar bug in DB.Get by changing levelIter.SeekGE to not return
the range deletion boundary key if the user-key portion is the key being
searched for. This avoids loading range tombstones from a table which
cannot possibly contain the search key,