Add support for prefix bloom filters
Regular bloom filtering adds a hash of the full user key (i.e. without
the internal seqno) into the bloom filter. Prefix bloom filtering
extracts a prefix from the input key instead. In particular, this allows
for the creation of iterators that can use bloom filters assuming one is
willing to restrict them to keys that share a common prefix, which is
common when implementing an MVCC scheme on top of the datastore (in
which case the "prefix" is the user key without the MVCC version).

This change stops short of actually allowing iterators to take advantage
of the prefix bloom filters. Since iterators are used even for point Get
operations, bloom filters won't currently be used at all for reading
data (outside of testing); this is left as a follow-up change.

Everything else is set up, though: the `Comparer` now has a `Split`
member which can be set to enable prefix extraction; SSTs will record
the extracted prefixes.

Full-key bloom filtering is still possible, though to keep things simple
at this point it is not possible to enable both prefix and full key
bloom filtering simultaneously. (It's unclear to the author whether that
is ever useful).