record: fix LogWriter's chunk alignment
Fixes #36.

This commit fixes two bugs in the record.LogWriter and adds tests for the
LogWriter alongside the record.Writer.

The first bug is that a record of length zero were skipped over and not
written. I'm not sure if this behaviour is ever exercised in practice,
but it is exercised by the tests for Writer so fixing it makes testing
simpler.

The second bug is the cause of #36: each chunk header is 7 bytes long,
and no chunk can span block boundaries. When there is insufficient space
left in a block to write an entire record, a "starting" chunk is written
and we move onto the next block to write the rest of it. If there are
exactly seven bytes available in the block, the original Writer would
write the seven byte header despite not having any room to write any
body bytes, while the LogWriter would omit them. This manifested as the
Reader thinking that it was seeing a zeroed out block.  This commit
brings LogWriter's behaviour in line with Writer does, and what the
Reader expects.

I've also made a change suggested by @petermattis to error out in the
case of a zero-block instead of attempting to recover. This required
removing a test, and was also masking bug #2 in tests. The log should
probably be truncated here and a log message generated stating that some
data was likely lost (but we have recovered to a consistent point in
time).