Fix iterator bug when range tombstone ends after its sstable
A range tombstone's end key may extend past its containing sstable's
upper bound. In that case, the portion of the range tombstone outside
the sstable's boundaries does not necessarily cover keys in lower levels.
The iterator optimization for seeking past keys covered by a range
tombstone in lower levels should take this into account and limit its
seek to the sstable's upper bound.

The approach taken here is to maintain the sstable upper bounds in
`mergingIter` for the partitioned levels. `levelIter` holds a pointer to
the upper bound for its level and updates it when the current sstable
changes. `mergingIter` then uses this info to prevent seeking too far in
lower levels upon encountering a range tombstone.