Make skiplist test more deterministic
The "TestConcurrentOneKey" test was assuming at least one write finished
before one hundred reads finished, which somehow was not the case 0.01%+
of the runs. We can wait for one write to finish at the beginning to
make sure this is the case. The test also runs in the old mode
(`startEmpty == false`) to prevent loss of coverage, except now it will
not require any writes to have completed in that mode.