Truncate range tombstones to atomic compaction unit boundaries
During compaction, truncate range tombstones at read time to atomic
compaction unit boundaries. This is accomplished by wrapping the
`rangeDelIter` used during compaction with a
`truncatedRangeDelIter`. This enforces the invariant that the boundaries
of the output of a compaction cannot be larger than the boundaries of
the input. It also enforces a tighter invariant regarding range
tombstones not enlarging beyond the boundaries of their atomic
compaction unit during compaction. This is necessary to prevent a range
tombstone from expanding after a compaction to cover a newer key at a
lower level in the LSM.

Fixes #26
Closes #27