Fix usage of syncingFile for sstables
The `sstable/Writer.bufWriter` was being initialized to the passed in
file, not our `syncingFile`.

The `rateLimitedFile` used in `DB.writeLevel0Table` was preventing the
usage of the `sync_file_range` code path on Linux. Recent
experimentation with rate limiting flushes has indicated that rate
limiting the file IO isn't the direction I want to go anyway.

Change `syncingFile.Sync` to use `fdatasync` on Linux. We don't need
file metadata updated when we sync.