Avoid holding DB.mu during log writes
Change `commitPipeline` to use a separate mutex rather than `DB.mu` for
serializing writing to the WAL. The use of `DB.mu` was a potential
performance problem because `DB.mu` is used for reads and holding it
during a log write meant that it was potentially held during disk
IO. There was also a super subtle bug here: `DB.makeRoomForWrite`
releases and then reacquires `DB.mu`. This behavior is necessary when
rotating the log, but the effect was that all of the careful ordering
guarantees in `commitePipeline` could be violated.

Fixes #44