Acquire state for reading without using DB.mu
Add a `readState` struct which holds the current version and memtable
slice. A `DB.readState` pointer is updated whenever the current version
or memtable slice changes.

name                    time/op
ReadState/updates=0-8   49.7ns ± 1%
ReadState/updates=10-8  89.9ns ± 2%
ReadState/updates=50-8   115ns ± 1%

Note that the `updates={10,50}` benchmarks are unrealistic. It is
extremely unlikely that 10% or 50% of the operations in a real workload
will be calls to `updateReadStateLocked`.

Fixes #45