Add support for the "recyclable" log record format
The recyclable log record format associated a log file number with each
record in a file which allows a log file to be reused safely without
truncating the file. This is an important performance optimization in
itself and a prerequisite for writing the WAL using direct IO.

Changed `record.LogWriter` to always use the recyclable log record
format which in turn changes Pebble to use the recyclable format for the
WAL. `record.Writer` (used to write the `MANIFEST`) continues to use the
legacy log record format. `record.Reader` can now read either format and
will return an error when a record with a log number does not match the
supplied log number.

Took the opportunity to unexport some methods that were not being used
outside of the `internal/record` package such as `Reader.SeekRecord` and
`LogWriter.Flush`.

Actual reuse of WAL files will be implemented in a future change.

See #41