Zero seqnums in sstables where possible
Zero seqnums in sstables in order to improve sstable compression and
enable an optimization to avoid a user-key comparison during forward
iteration. The seqnum for an entry can be zeroed if the key does not
exist at lower levels in the tree and if the entry is in the last
snapshot stripe. Rather than performing the check on a per-key basis, a
single check is done to see if the bounds of a compaction overlaps with
any sstables at lower-levels in the tree. This is slightly
pessimistic (it disallows zeroing of some keys when they could be
zeroed), but fast.

Forward iteration normally requires a user-key comparison to determine
if we've moved to the next user-key. We can skip this comparison if the
seqnum for an entry is 0 as that is the oldest seqnum for a particular
user-key which guarantees the next entry is for a different
user-key. This produces a ~5% speedup in forward iteration performance
when iterating over data contained in the bottommost sstables. Note the
rule of thumb is that ~90% of data in an LSM is in the bottommost level.

Fixes #87