Fix the trimming of the upper-bound for sstable iteration
Previously we were trimming the upper-bound comparison of the
upper-bound was >= the index key. As the comments already note: The
separator (the index key) between blocks i and i+1 is a key that is >=
every key in block i and is < every key i block i+1.

The existing `TestReader*` tests caught this problem when stressed,
failing with the error:

  table_test.go:376: expected 777, but found 778.

This occurred whenever the upper-bound `king` was selected, as that is
both an index key and the entry in a block.