Add basic support for "log data"
Log data is a batch entry that gets written to the WAL, but is not
added to the database state. RocksDB uses log data to implement 2PC. It
is also useful for performance testing the WAL in isolation from other
the other parts of the system.

Add `--disable-wal` and `--wal-only` flags to the `pebble` tool.