Rework DB.deleteObsoleteFiles
Previously, deleteObsoleteFiles would perform a directory listing to
find obsolete files. The directory listing is fairly expensive,
especially as the number of tables gets large. Switch to an approach
that is similar to RocksDB, where in steady state we keep track of
obsolete logs and tables as memtables are flushed and versions are
unreferenced.

Fix a bug in handling of `versionSet.logNumber`. Previously that field
was being bumped to the most recent log file number any time a flush was
performed, but we can only bump it to the most recent log file that
hasn't been flushed.

Fixes #96