Add support for capping the size of the MANIFEST
The MANIFEST is a log (in `internal/record` format) containing a
sequence of version edits. A version edit is created whenever a flush or
compaction occurs. The actual current version is obtained by replaying
the log at startup (the current version is also maintained in memory, of
course). In order to avoid the MANIFEST growing excessively large, it is
"rolled" whenever its size exceeds
`Options.MaxManifestFileSize` (default 128 MB). The new MANIFEST is
completely self-contained with the version entry being a snapshot of the
entire state (essentially a delta from the empty state).

Re-jiggered the way the MANIFEST is initially created which resulted in
slightly different file numbering at database creation: the first log
now gets file number 2, and the manifest file number 3.

Fixes #42