Use largest optype when overwriting file's smallest key
This prevents the non-overlapping invariant from failing in cases where
the file's smallest key is overwritten according to the previous file's
largest key. We should use the largest optype (value zero since optypes
have descending ordering) to prevent overlap.

After the fix in #118, this only happens in contrived scenarios.
That's because we were already using seqnum zero (the largest seqnum)
when overwriting smallest key. For overlap to occur, then, the previous
file's largest would have to be a point key with seqnum zero, and the
current output file would require a range tombstone extending into the
previous file. If a range tombstone extends into the previous file,
though, that file will contain the same range tombstone and extend its
largest to the current file's first point key. That first point key must
come after all point keys in the previous file since the previous file's
last point key had seqnum zero and there cannot be two of the same keys
with seqnum zero. So the situation described is impossible as it caused
a contradiction.