Metamorphic test
Randomly generate a sequence of operations, and then replay this
sequence repeatedly with random options, flushes, and compactions,
verifying the test history is the same on each run.

The output from running a sequence of operations is recored in a
`history`. The format of this output looks similar to Go code. The
operations themselves look like method invocations. The results from an
operation are recorded in comments. Event logs from Pebble are also
recorded in comments. This output can be parsed to regenerate the
sequence of ops that led to a failure or discrepancy between runs.