Bucket filling / draining simulation
Simulation of filling and draining a bucket in chunks with
bi-directional flow control. If draining is proceeding faster than
filling, then draining is slowed down (but kept at a minimum rate so it
doesn't stop completely). If filling is proceeding faster than draining,
it is slowed down.

The simulation is meant to represent filling and draining of the
memtable. Each memtable is a chunk and draining (flushing) of a memtable
cannot proceed until the memtable is full. The simulation periodically
adjusts the rate of filling and demonstrates the nimble reaction of the
flow control mechanism.

See #7