Sync the DB dir whenever we manipulate directory entries
Whenever an sstable is created (due to flush or compaction) or a WAL or
MANIFEST is created, we need to sync the parent directory in order to
ensure that the directory update is persisted to disk.

Set O_CLOEXEC and all of the files created and opened by Pebble, to
match RocksDB behavior.

Fixes #125