Support storing WALs in a separate directory
Add `Options.WALDir` which specifies the directory to store write-ahead
logs in. This allows write-ahead logs to be stored on faster disk than
sstables. For example, write-ahead logs could be stored on
battery-backed SSD, while sstables are stored on spinning disk. Or
write-ahead logs could be stored on an Optane drive, while sstables are
stored on SSDs. If `Options.WALDir` is empty, write-ahead logs are
stored in the primary data directory.

Fixes #65