Adjust dynamic-base-level heuristic
Previously, the target size of the base level was computed as
`max(L1MaxBytes,L0Size)`. This had the negative effect of allowing the
target size for the base level to grow dramatically if L0 was growing
dramatically, which would prevent the dynamic-base-level heuristic from
adding additional levels to the LSM.

Rename `Options.L1MaxBytes` to `Options.LBaseMaxBytes`.