Change sstable.NewWriter to take a writeCloseSyncer
The writeCloseSyncer is a combination of the `io.WriteCloser` interface
and a `Sync()` method. This interface is much smaller than `vfs.File`
and makes it easier to generate sstables without implementing the full
`vfs.File` interface.

Moved wrapping of `vfs.File` with `NewSyncingFile` from
`sstable.NewWriter` to its callers.

Fixes #140