Replace db package with internal/base
The `pebble/db` package name is somewhat problematic as it consumes the
name `db` or requires the user to rename that package during import to
something like `pebble_db`. The reason that package existed was to
provide a base location for types that need to be shared between the
`pebble` package and various sub-packages such as
`pebble/sstable`. Rather than a shared package that is also available to
the user, the types are now defined in `pebble/internal/base` and
exposed in `pebble` and `pebble/sstable` via type aliases.