Convert v1 to v2 range-del blocks on the fly
RocksDB generates range-del blocks that contain range tombstones that
are not fragmented and sorted. Pebble generates range-del blocks
containing tombstones that are fragmented and sorted. The fragmented and
sorted range-del blocks are suitable for serving from directly, while
the unfragmented/unsorted blocks are not. We now transform RocksDB
format range-del blocks into Pebble format blocks on the fly.

The transformation is done via a hook that allows passing a block
through a function after it has been read from disk but before it is
added to the cache.

Fixes #68