Provide key stability across blockIter positioning calls
Provide stability for the keys returned from a blockIter across
positioning calls (Seek*, First, Last, Prev, Next) when the block has a
restart interval of 1. This is achieved by noticing when a key does not
share any bytes with the previous key and changing the returned key to
point directly into the block instead of the key buffer. This behavior
was already being expected by `rangedel.Truncate` and by forthcoming
code to transform v1 range-del blocks to v2 format on the fly.