Simplify and correct ycsb.nextReadKey
Do not hash the key returned by the key distrbution, and remove the
mod. The mod is not necessary because the range of keys returned by the
key distribution is only increased as the acknowledged sequence set base
increases. The hash is not necessary because we're already hashing in
ycsb.makeKey.