Fatal on MANIFEST write errors
When a MANIFEST write fails, recovery is difficult because it is unclear
what has made it to disk and what hasn't. Rather than trying to perfom
inline recovery to a good state, we fatal and rely on the standard
recovery mechanism run when a database is open. In particular, that
mechanism generates a new MANIFEST and ensures it is synced.

Fixes #153