Fix a concurrency bottelneck in tableCache
The locking mechanism in tableCacheNode for retrieving the associated
sstable.Reader was showing up as a significant bottleneck. This was
clearly seen when running a read-only ycsb workload:

./pebble ycsb bench \
  --workload scan=100 \
  --duration 10s \
  --initial-keys 0 \
  --prepopulated-keys 10000000 \
  --keys uniform \
  --concurrency 10 \
  --scans 1

____optype__elapsed_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
  scan_100    10.0s        1037672       103771.7      0.1      0.1      0.1      0.2     16.8

With `--concurrency 1`, the performance was nearly identical. For a
workload like this, we expect the performance to increase almost
linearly with the number of CPUs. The blocking profile showed that the
tableCacheNode channel was the source of all of the blocking
contention. The previous mechanism stored the sstable.Reader inside of
the channel, and every call to tableCache.newIters would read the reader
from the channel, then put the reader back on the channel. This has now
been replaced with an unbuffered channel which is used to signal when
loading of the sstable.Reader is complete. The reader itself is stored
in the tableCache.reader field. The new results for the workload above
are:

____optype__elapsed_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
  scan_100    10.0s        3807782       380860.7      0.0      0.0      0.1      0.2     11.0

Not quite 4x speedup (on my 4 core laptop), but pretty close.