Add a hit buffer to batch tableCacheShard LRU updates
Updating the tableCacheShard LRU lists requires mutually exclusive
access to the list. Grabbing and releasing tableCacheShard.mu for each
access shows up prominently on blocking profiles. There is a known
technique for avoiding this overhead: record the access in a per-thread
data structure and then batch apply the buffered accesses when the
buffer is full. For a highly concurrent scan workload, this results in a
significant improvement in both throughput and stability of the
throughput numbers. Here are numbers when running with 100 concurrent
scanners before this change:

____optype__elapsed_____ops(total)___ops/sec(cum)
  scan_100   120.0s      109476226       912303.4

After:

____optype__elapsed_____ops(total)___ops/sec(cum)
  scan_100   120.0s      151736009      1264384.6

That's a 39% speedup. Even better, the instantaneous numbers during the
run were stable.

See #11