Add lightweight manual memory management for the block cache
Add support for lightweight manual memmory management to the block
cache. Memory is still allocated using the Go runtime, yet we track
blocks that are either in the cache, or still in use by an
iterator. When the final reference to a block is released, the byte
slice is added to a per-thread list of recently freed
blocks. Allocations for block memory are first satisfied from this cache
before falling back to the Go allocator.

The allocation cache groups allocation sizes into 1024-byte
classes. There is some space wastage in this setup which we could reduce
with finer granularity classes such as those used in the Go runtime.

Experimentation shows that 99% of allocation sizes for blocks cluster
within a single size class: the configured target block size. In a
workload where the working set exceeds the block cache size, the
allocation cache effectively eliminates 99% of the allocations that
would occur as blocks are evicted from the cache and then another block
is added.

Fixes #31