Fix a "memory leak" in caching entries during reverse iteration
During reverse iteration, the cache of entries to the previous restart
point was holding a pointer into `blockIter.cachedBuf`. That buffer
started size at 0 and then increased on every entry added. If the buffer
had to be resized, existing cached entries would still point to the old
buffer. The result was not incorrect, but it used more memory than
expected.