Fix block flushing heuristic during sstable construction
The block flushing heuristic in `sstable/Writer.maybeFlush` had a small
bug which was previously partially worked around by another bug. The
small bug was due to the usage of `&` instead of `%`. Fixing that
allowed the TODO to be fixed, which in turn fixed the discrepancy
between the sstables generated by Pebble and RocksDB.

Changed the sstable test fixtures to use a 2048 target block size. This
was sufficient to reveal the problem as it stress the block flushing
heuristic more.

Fixes #192