Fix ingestSortAndVerify overlapping check
Previously, the end key of a range deletion tombstone was considered
exclusive for the purposes of deletion, but considered inclusive when
checking if two SSTables overlap. For example, an SSTable with a range
deletion tombstone [a, b) would be considered overlapping with an
SSTable with a range deletion tombstone [b, c). This commit fixes this
check.