Add introspection commands to pebble tool
Move the `pebble` benchmarks into a new `bench` subcommand. Add
skeletons for `db`, `manifest`, `sstable`, and `wal` introspection
commands.

Implement `sstable layout` and `sstable properties`. The `sstable
layout` command shows the layout of blocks and records in an
sstable. The `properties` command shows the contents of the properties
block.

See #202