cache: add a dbNum argument to cache methods
Allow a cache to be shared by multiple Pebble instances by adding a
`dbNum` argument to the cache methods which take a `fileNum`. Plumb this
through everywhere, which is the majority of changes.

Provide a global `dbNum` allocator. This "uniquifier" only needs to
provide uniqueness within a single process.

Change the default options to create an 8 MB cache rather than not
caching at all. This matches the RocksDB behavior.

Fixes #206