Fix a triplet of synchronization bugs
The previous reworking of the commitQueue implementation accidentally
omitted an important invariant: `commitQueue.dequeue()` should only
return "applied" batches. The lack of this check was allowing the
visible sequence number to be ratcheted for unapplied batches. Added an
assertion that this invariant is not violated.

The change to not hold `DB.mu` during log writes failed to adequately
protected `DB.mu.log.LogWriter` allowing that field to be modified
concurrently with a call to `LogWriter.SyncRecord`. The concurrent
modification was occurring because `DB.mu` no longer protects that
field, yet some callers of `DB.makeRoomForWrite` were not
synchronized. Specifically, the calls in
`DB.{Compact,Flush,AsyncFlush}`.

The change to not use `DB.mu` for commit pipeline synchronization also
led to a new synchronization requirement on `versionSet.logSeqNum`. That
field is incremented while holding `commitPipeline.mu`, but read while
holding `DB.mu`. Neither mutex protects it! The solution is to increment
and read that field atomically.

Added a test which stresses concurrent commit, flush, and compact
operations. All of the above problems were easily visible by running
this test under the race detector.