sstable: fix curOffset calculation at end of compressed block
The previous calculation of `curOffset` at the end of a compressed
block caused the value to be larger than the end offset of the block on
disk. This is problematic because the next block would then set
`curOffset` to a smaller value. This in turn would lead to
`bytesIterated` being incremented by a very large value (a negative
value stuffed in a `uint64`) which would in turn cause the compaction
pacer to enter a loop which would take ~forever to terminate.