Verify comparer and merger name on sstable open
The sstable properties contain the comparer and merger names configured
when the sstable was written. When opening an sstable, verify that the
comparer and merger names match the ones specified in the options.

Switch to using the comparers, filters, and mergers map when opening an
sstable. This allows tools to open sstables where the sstable can have
one of multiple comparers or mergers.