Unbreak the build on freebsd (and probably netbsd)
Note: linux also supports `posix_fallocate(2)` however it may be worh
noticing the difference `syscall.Fallocate` (the linux-specific
syscall `fallocate(2)`): the former is specified to pre-initialize the
space with zeroes, whereas the latter leaves the space uninitialized.

There is a security/performance trade-off which needs to be analyzed
but is out of scope for this commit.

- freebsd: https://www.freebsd.org/cgi/man.cgi?query=posix_fallocate&apropos=0&sektion=0&manpath=FreeBSD+12.0-RELEASE+and+Ports&arch=default&format=html
- netbsd: https://netbsd.gw.com/cgi-bin/man-cgi?posix_fallocate++NetBSD-current
- openbsd: not supported (nor is there a `fallocate`)
- dragonfly: not supported http://gitweb.dragonflybsd.org/dragonfly.git/blob_plain/HEAD:/sys/sys/fcntl.h
- solaris: could not find anything