Add TableCreated event
Add an event for table creation. This event fires immediately after the
sstable file is created which means the file will be empty (nothing has
been written to it yet).

Refactor the write stall event tests to use the TableCreated event for
detecting when a table is created for flush vs compaction.