Restructure the tool commands
Rather than global commands, provide the tool commands via structs. This
will make it more natural for configuring options associated with the
commands, such as registering pretty printers or additional comparers or
mergers.

Add support for specifying start and end keys for `sstable scan` in
hex. Add support for specifying the format of keys and values for
`sstable scan`.