batch: Don't increment count on LogData(), add misc helpers
RocksDB batches contain a count field in their byte representation
which is usually NOT incremented for any operation of kind LogData
that exists in the WAL but isn't applied to the memtable. As
Cockroach moves to using pebble.Batch instead of its own batch builder,
reconcile this difference so RocksDB doesn't error out on pebble.Batch
produced batches.

Also add some exported helper methods to Batch - one of them being
SetRepr to set a specified byte slice as the underlying binary
representation of the batch. Others are {Set,Merge,Delete}Encodeable
which are similar to their non-Encodeable variants except they accept
an Encodeable object instead of a byte slice. Useful for passing in
objects that Pebble doesn't know how to encode (eg. MVCCKeys in
Cockroach).

Also exports Reset(), useful for reusing the same batch without
allocating a whole new one. And export Count() to retrieve the count
stored inside the batch's binary representation.