Various EventListener improvements
Note a `TableCreated` event during sstable ingestion. Previously this
was omitted.

Add `{Flush,Compaction}Info.Done` so that the `{Flush,Compaction}End`
events can accurately distinguish via the `*Info` between a completed
flush/compaction and one that produced no output.

Call `{Flush,Compaction}End` after the manifest has been updated so that
the info will include any error in applying the flush/compaction.

Change the format of `Version.{Pretty,DebugString}` to be closer to the
output of `manifest dump`.

Fix a bug in which we were failing to cleanup the flush/compaction
output if there was an error applying the results to the manifest.