Fix incorrect manifest.CheckOrdering checks
The `manifest.CheckOrdering` checks were overly strict, rejecting valid
L0 orderings. In particular, the checks were rejecting orderings that
can occur after ingestion. An sstable ingested into L0 can be given a
global seqnum which lies in the middle of another L0 sstable. This is ok
because during ingestion we checked that the ingested sstable does not
overlap in the key range of the other sstable.