Fix and test Fragmenter.FlushTo
Fix bugs in `Fragmenter.FlushTo`: it could emit tombstones after the
specified key and cause the creation of empty tombstones which would
later be emitted. Document that `FlushTo` does not truncate tombstones
as truncating tombstones to sstable boundaries has to be done at read
time for correctness, not at compaction time. (The specific scenario is
described in comments in the code).

Add testing for `FlushTo`, and add testing of the invariant checking for
proper use of `Fragmenter`.