Run tests on Windows and OS X during CI
Change memFS to use "/" as a path separator rather than
`os.PathSeparator`. This allows datadriven tests to hard code "/" in
their expected test data, and not have to handle "\\" on Windows. Added
`vfs.FS.{PathBase,PathJoin}` routines for abstracting the path separator
handling. This required mechanical changes to a bunch of code.

Changed two "package comments" (in `cache` and `sstable`) which were
causing the linter to complain on Windows. I'm not sure exactly why, but
changing the comment style fixes things.

One test is skipped on Windows: `internal/rate.TestLongRunningQPS` is
timing dependent and fails reliably on Windows due to the resolution of
the timer.