Run tests on Windows and OS X during CI
Change memFS to use "/" as a path separator rather than
`os.PathSeparator`. This allows datadriven tests to hard code "/" in
their expected test data, and not have to handle "\\" on Windows. Added
`vfs.FS.{PathBase,PathJoin}` routines for abstracting the path separator
handling. This required mechanical changes to a bunch of code.

Changed two "package comments" (in `cache` and `sstable`) which were
causing the linter to complain on Windows. I'm not sure exactly why, but
changing the comment style fixes things.

Two tests are skipped on Windows. The `internal/rate.TestLongRunningQPS`
test because it appears to be timing dependent. The other failure is
more disconcerting. `vfs.TestLock` should succeed, but is failing to
acquire the lock in the parent process indicating that it is already in
use. Someone with more Windows knowledge will need to investigate.