internal/record: add support for propagating sync errors
Add support for propagating sync errors when syncing the WAL. Previously
such errors caused the commit pipeline to halt indefinitely waiting for
the sync to occur successfully in order for a batch to commit. A
separate change adjusts the order of filesystem operations which caused
`TestErrors` to reveal this error handling gap.