db: improve DB creation
Improve DB creation so that it uses the normal paths for writing to the
MANIFEST rather than custom code. This had the side-effect of avoiding
consuming an extra file number for the MANIFEST at creation
time (previously we were creating MANIFEST-000001, then replaying 0 log
files and creating MANIFEST-000003). This change resulted in number of
changes to test output as expected file numbers shifted.