fix "queue is full" panic
The `commitPipeline` and `record.LogWriter` coordinate in the amount of
concurrency allowed for `LogWriter.SyncRecord` operations. The contract
is that `SyncRecord` can only be called by `SyncConcurrency` goroutines
concurrently. This was failing to take into account the implicit
`SyncRecord` performed by `LogWriter.Close`. The off-by-one could lead
to a "queue is full" panic from the `LogWriter` code.

Fixes #301