fix rangedel.Fragmenter panic
Fix a panic in rangedel.Fragmenter: `flush-to key < flushed key`. The
problem was that `compactionIter` was adding tombstones to the
fragmenter too early. In particular, we hit this panic by adding
tombstones that had a larger start key than the current key. If the
compaction logic decided to switch to the next sstable, we'd flush up to
the current key which is less than the start key of the most recently
added tombstone.

Fixes #317