db: rework how cache IDs are allocated
Avoid a common source of errors which made it easy to accidentally get
cache collisions by preventing the use of the zero cache ID. Renamed
`dbNum` to `cacheID` to make it clear that the cache can be used without
a DB.

Added `Cache.NewID()` for allocating cache IDs. Changed
`sstable.NewReader` to allocate a new `cacheID` via `Cache.NewID()` if 0
is passed as an argument.