avoid (re-)computing Batch.memTableSize when it won't be used
CockroachDB uses pebble.Batch for building batches. This avoids a
performance a regression due to the recomputation of
`Batch.memTableSize` whenever `Batch.Apply` and `Batch.SetRepr` are
called. `Batch.memTableSize` is only used by Pebble when committing a
batch. Rather than computing it up front, we can delay computation if
the `Batch` is not associated with a `DB`.