sstable: NewReader cleanup on error
Changed `NewReader()` to return either a Reader on success, or an
error on failure, but never both. This should be more conventional.

Also introduced `vfs.NewMemFile()` on the pretense of it being useful
for this PR's unit test. Though really it is needed to simplify a PR in
Cockroach repo.