sstable: add sstable.{,Table}Options
Rather than defining `sstable.Options` as an alias for `base.Options`,
and `sstable.TableOptions` as an alias for `base.LevelOptions`, define
actual structs that contain only the options necessary for reading and
writing sstables. The options are now always passed by value to
`New{Reader,Writer}` which avoids unintentional sharing of options
between readers and writers that was a frequent source of race
conditions.

Changed the default cache size for `sstable.Options` to be
empty. Previously it was inheriting the default from `pebble.Options`
which is an 8MB cache.