reserve space for memtables from the cache
Add `memTable.readerRefs` and methods to add and remove reader
references when memtables are referenced by a `readState.memtables` and
when they are removed from `DB.mu.mem.queue`. Track the total reserved
bytes by memtables in `DB.memTableReserved` and check that this count
falls to zero when a DB is closed.

Fixes #30