truncate sstable boundaries to the compaction input boundaries
When determining the boundaries for an sstable, we need to truncate the
boundaries to the boundaries of the input. Failing to do this was
allowing the lower bound of the first sstable in a compaction to expand
beyond the lower bound of the input when the sstable contain a range
tombstone.

Fixes #337