add missing compaction.expandInputs calls
Add missing `compaction.expandInputs` calls when the inputs for a
compaction are being determined. `expandInputs` expands the input
sstables for a level to include adjacent files up to a "clean
cut". Consider the scenario with 2 L6 sstables: [a#4,1-b#3,1] and
[b#2,1-c#1,1]. Before this change, the first sstable could be compacted
independently of the right sstable, leading to an output sstable
[a#0,1-b#0,1] (due to seqnum zeroing) which then overlaps with the
second sstable.

Rename `compaction.setupOtherInputs` to `compaction.setupInputs`. It
wasn't clear what the "other" was in this case. Presumably the output
level inputs, but `setupOtherInputs` was also affect the start level
inputs.

See #337