explicitly disable go module support for {test*,stress*,generate}
We're using go modules to vendor, but `GO111MODULE=auto` now enables
modules when running within `GOPATH` under go1.13. Explicitly disable
the usage of the module cache so that we can use the vendored source. We
could also specify `-mod=vendor` under go1.13, but that flag doesn't
work the same under go1.12.

Removed the `bench` target. This was useful at one time, but with the
growth of the code base it is better/easier to run `go test -v -run -
-bench <test>` from the command line.