internal/record: wait for flush-loop to terminate in LogWriter.Close
`LogWriter.Close()` was previously failing to wait for the flush-loop to
terminate which led to a rare race condition in the `internal/record`
tests where a buffer shared between a writer and a reader could be
written to by the writer's flush-loop after the writer had been
closed.

Added additional synchronization to allow waiting for the flush-loop to
terminate after `Close` has signalled for it to close. Tweaked the way
the final sync is performed, so that it is done directly in `Close()`
rather than asking the sync loop to perform it. The problem with asking
the sync loop to perform the close was related to the above race: we
could ask for a sync, but miss flushing of the data in a partially
filled block.

Fixes #356