ingest: Move file paths into FileMetadata
Currently we assume that the paths slice is in the same ordering as
associated fileMetadatas, but we sort the latter without modifying
the former. This leads to cases where the two mismatch, resulting
in manifests being out of sync with SSTs on disk.

This change moves the path inside the FileMetadata itself so the
two don't go out of sync.

Discovered when cockroach snapshots wounded up breaking ranges
they were applied on.