ingest: Keep paths and fileMetas in sync when sorting
Currently we assume that the paths slice is in the same ordering as
associated fileMetadatas, but we sort the latter without modifying
the former. This leads to cases where the two mismatch, resulting
in manifests being out of sync with SSTs on disk.

This change ensures the file paths slice as well as the metas
slice is swapped together when sorting.

Discovered when cockroach snapshots wounded up breaking ranges
they were applied on.