fix resetting of Batch.count for large batches
We were failing to clear `Batch.count` for large batches which was
leading to some super strange `make stress` errors. This turned out to
be extremely onerous to debug which resulted in most of the code changes
here (the fix is the small change to `Batch.Reset`).

The `make stress` failure was a panic comming out of `syncQueue.pop()`
because the `WaitGroup` had a negative counter. How did an incorrect
`Batch.count` result in a negative `WaitGroup` counter? The incorrect
`Batch.count` was causing `memTable.apply` to panic. The panic there was
propagating up through `DB.LogData` which had a `defer
Batch.release()`. Defers get run on the panic return, so we were calling
`Batch.release()` on a batch still sitting on the
`syncQueue`. Internally, `Batch.release()` resets `Batch.commit` (the
`WaitGroup` passed to `LogWriter.SyncRecord`). When the `LogWriter`
eventually performed the sync, it called `WaitGroup.Done()` on what was
essentially a zero-initialized object. Boom!

To mitigate such debugging odysseys in the future, most (but not all)
panics in and near `commitPipeline.Commit` have been replaced with error
returns. `DB.Apply`, the singular caller of `commitPipeline.Commit`, now
calls `Logger.Fatalf` on error. This is preferrable to the panic because
we will get a clean termination and not an obscure one.

Lastly, the `defer Batch.release()` pattern has been removed and now we
only release batches when they successfully apply. Unsuccessful
application will result in the batch being left for GC.

Fixes #346